+++ semcore/dropdown/src/AbstractDropdown.tsx
@@ -68,20 +68,21 @@ export abstract class AbstractDropdown extends Component<AbstractDDProps, {}, {}
 
     if (interaction === 'none') return false;
 
-    e.preventDefault();
-    e.stopPropagation();
-    this.handlers.visible(true);
-
     setTimeout(() => {
-      const highlightedIndex = this.asProps.highlightedIndex ?? 0;
-      const element = this.itemRefs[highlightedIndex];
-      element?.focus();
-      if (this.role === 'menu') {
-        this.handlers.highlightedIndex(highlightedIndex);
+      const { visible, inlineActions } = this.asProps;
+      if (visible || inlineActions) {
+        this.afterOpenPopper();
       }
-    }, 0);
+    }, 200); // because first will be executed onClick handler in popper
+  };
 
-    return false;
+  afterOpenPopper = () => {
+    const highlightedIndex = this.asProps.highlightedIndex ?? 0;
+    const element = this.itemRefs[highlightedIndex];
+    element?.focus();
+    if (this.role === 'menu') {
+      this.handlers.highlightedIndex(highlightedIndex);
+    }
   };
 
   getTriggerProps() {
@@ -180,7 +181,7 @@ export abstract class AbstractDropdown extends Component<AbstractDDProps, {}, {}
     const itemsLastIndex = this.itemProps.length - 1;
     const selectedIndex = this.itemProps.findIndex((item) => item.selected);
 
-    if (itemsLastIndex < 0) return 0;
+    if (itemsLastIndex < 0) return -1;
 
     let innerHighlightedIndex: number;
 
@@ -206,7 +207,7 @@ export abstract class AbstractDropdown extends Component<AbstractDDProps, {}, {}
     if (this.itemProps[newIndex]?.disabled) {
       return this.getHighlightedIndex(amount < 0 ? amount - 1 : amount + 1);
     } else if (!this.itemProps[newIndex]) {
-      return 0;
+      return -1;
     } else {
       return newIndex;
     }
@@ -265,6 +266,10 @@ export abstract class AbstractDropdown extends Component<AbstractDDProps, {}, {}
       ['ArrowDown', 'ArrowUp', 'Enter', ' '].includes(e.key) &&
       e.currentTarget.getAttribute('role') !== this.childRole
     ) {
+      if (['ArrowDown', 'ArrowUp'].includes(e.key)) {
+        this.handlers.visible(true);
+      }
+
       this.handleClickTrigger(e);
     }
   }
@@ -302,7 +307,11 @@ export abstract class AbstractDropdown extends Component<AbstractDDProps, {}, {}
       }
       case ' ':
       case 'Enter':
-        if (this.highlightedItemRef.current && highlightedIndex !== null) {
+        if (
+          this.highlightedItemRef.current &&
+          highlightedIndex !== null &&
+          !this.itemProps[highlightedIndex].disabled
+        ) {
           e.stopPropagation();
           e.preventDefault();
           this.highlightedItemRef.current.click();
@@ -313,9 +322,7 @@ export abstract class AbstractDropdown extends Component<AbstractDDProps, {}, {}
 
     if (amount !== null) {
       const newHighlightedIndex = this.getHighlightedIndex(amount);
-      if (newHighlightedIndex !== undefined && this.role === 'menu') {
-        this.itemRefs[newHighlightedIndex]?.focus();
-      }
+
       if (
         this.role === 'listbox' &&
         this.triggerRef.current &&
@@ -323,7 +330,13 @@ export abstract class AbstractDropdown extends Component<AbstractDDProps, {}, {}
       ) {
         this.focusTrigger();
       }
-      this.handlers.highlightedIndex(newHighlightedIndex, e);
+
+      if (newHighlightedIndex !== -1) {
+        this.handlers.highlightedIndex(newHighlightedIndex, e);
+        if (this.role === 'menu') {
+          this.itemRefs[newHighlightedIndex]?.focus();
+        }
+      }
 
       e.preventDefault();
       e.stopPropagation();