+++ semcore/select/__tests__/select.browser-test.tsx
@@ -1,8 +1,8 @@
 import { expect, test } from '@semcore/testing-utils/playwright';
 import { e2eStandToHtml } from '@semcore/testing-utils/e2e-stand';
 
-test.describe('Select', () => {
-  test('Traps focus', async ({ page }) => {
+test.describe('Options Filtering', () => {
+  test('Traps focus on input', async ({ page }) => {
     const standPath = 'stories/components/select/docs/examples/options_filtering.tsx';
     const htmlContent = await e2eStandToHtml(standPath, 'en');
 
@@ -15,7 +15,7 @@ test.describe('Select', () => {
 
     await expect(page).toHaveScreenshot();
   });
-  test('Returns focus', async ({ page }) => {
+  test('Closes and Returns focus by esc', async ({ page }) => {
     const standPath = 'stories/components/select/docs/examples/options_filtering.tsx';
     const htmlContent = await e2eStandToHtml(standPath, 'en');
 
@@ -33,7 +33,44 @@ test.describe('Select', () => {
 
     await expect(page).toHaveScreenshot();
   });
-  test('Traps focus and handles clicks', async ({ page }) => {
+  test('Closes by mouse click on trigger', async ({ page }) => {
+    const standPath = 'stories/components/select/docs/examples/options_filtering.tsx';
+    const htmlContent = await e2eStandToHtml(standPath, 'en');
+
+    await page.setContent(htmlContent);
+    const Trigger = page.locator(
+      '[data-ui-name="ButtonTrigger.Text"][placeholder="Select a fruit"]',
+    );
+    await Trigger.click();
+
+    await page.waitForSelector('input');
+    const inputLocaltor = await page.locator('input');
+
+    await Trigger.click();
+    await expect(inputLocaltor).toHaveCount(0);
+    await expect(inputLocaltor).not.toBeVisible();
+  });
+  test('Open select and choose option by keyboard', async ({ page }) => {
+    const standPath = 'stories/components/select/docs/examples/options_filtering.tsx';
+    const htmlContent = await e2eStandToHtml(standPath, 'en');
+
+    await page.setContent(htmlContent);
+
+    await page.keyboard.press('Tab');
+    await page.keyboard.press('Space');
+    await page.waitForSelector('input');
+    const inputLocaltor = await page.locator('input');
+
+    await new Promise((resolve) => setTimeout(resolve, 500));
+    await page.keyboard.press('ArrowDown');
+    await page.keyboard.press('ArrowDown');
+
+    await page.keyboard.press('Space');
+    await new Promise((resolve) => setTimeout(resolve, 500));
+    const triggerLocatior = await page.locator('[data-ui-name="Select.Trigger"]');
+    await expect(await triggerLocatior.textContent()).toBe('Banana');
+  });
+  test('Open select and choose option by mouse', async ({ page }) => {
     const standPath = 'stories/components/select/docs/examples/options_filtering.tsx';
     const htmlContent = await e2eStandToHtml(standPath, 'en');
 
@@ -43,9 +80,7 @@ test.describe('Select', () => {
     await page.keyboard.press('Space');
 
     await page.waitForSelector('input');
-
     const inputLocaltor = await page.locator('input');
-
     await new Promise((resolve) => setTimeout(resolve, 500));
 
     const optionBLocator = await page.locator('text=Banana');
@@ -63,6 +98,7 @@ test.describe('Select', () => {
 
     await expect(await triggerLocatior.textContent()).toBe('Banana');
   });
+
   test('Show hint for clear input button by keyboard', async ({ page }) => {
     const standPath = 'stories/components/select/docs/examples/options_filtering.tsx';
     const htmlContent = await e2eStandToHtml(standPath, 'en');
@@ -100,8 +136,10 @@ test.describe('Select', () => {
     await new Promise((resolve) => setTimeout(resolve, 500));
     await expect(page).toHaveScreenshot();
   });
+});
 
-  test('Show highlight the first option item in render function', async ({ page }) => {
+test.describe('Render function', () => {
+  test('Show highlight the first option item', async ({ page }) => {
     const standPath = 'stories/components/select/docs/examples/render_function.tsx';
     const htmlContent = await e2eStandToHtml(standPath, 'en');
 
@@ -116,5 +154,30 @@ test.describe('Select', () => {
     await page.keyboard.press('Enter');
 
     await expect(page).toHaveScreenshot();
+  });
+});
+
+test.describe('Focus interaction', () => {
+  test('Open By mouse', async ({ page }) => {
+    const standPath = 'stories/components/select/tests/examples/focus_interaction.tsx';
+    const htmlContent = await e2eStandToHtml(standPath, 'en');
+
+    await page.setContent(htmlContent);
+    const trigger = page.locator('button[placeholder = "Select option"]');
+    await trigger.click();
+    await expect(page.locator('[data-ui-name ="DropdownMenu.Popper"]')).toBeVisible();
+    await trigger.click();
+    await expect(page.locator('[data-ui-name ="DropdownMenu.Popper"]')).toBeVisible();
+  });
+
+  test('Open By keyboard', async ({ page }) => {
+    const standPath = 'stories/components/select/tests/examples/focus_interaction.tsx';
+    const htmlContent = await e2eStandToHtml(standPath, 'en');
+
+    await page.setContent(htmlContent);
+    await page.keyboard.press('Tab');
+    await expect(page.locator('[data-ui-name ="DropdownMenu.Popper"]')).toBeVisible();
+    await page.keyboard.press('Escape');
+    await expect(page.locator('[data-ui-name ="DropdownMenu.Popper"]')).not.toBeVisible();
   });
 });