+++ semcore/select/src/Select.jsx
@@ -143,7 +143,9 @@ class RootSelect extends AbstractDropdown {
       'aria-haspopup': isMenu ? 'listbox' : 'dialog',
       'aria-disabled': disabled ? 'true' : 'false',
       'aria-activedescendant':
-        visible && highlightedIndex !== null ? `igc-${uid}-option-${highlightedIndex}` : undefined,
+        visible && highlightedIndex !== null && this.itemRefs[highlightedIndex]
+          ? `igc-${uid}-option-${highlightedIndex}`
+          : undefined,
       empty: isEmptyValue(value),
       value,
       name,
@@ -156,6 +158,21 @@ class RootSelect extends AbstractDropdown {
       onClear: this.handlerClear,
       children: this.renderChildrenTrigger(value, options),
       getI18nText,
+
+      onBlur: () => {
+        // if popper is opened and we moved from the trigger in select - it means we moved on some controls in popper and should hide highlighted for the option
+        if (this.asProps.visible) {
+          this.prevHighlightedIndex = this.asProps.highlightedIndex;
+          this.handlers.highlightedIndex(null);
+        }
+      },
+      onFocus: () => {
+        // if popper is opened and we moved to the trigger in select - it means we moved from some controls in popper and should highlight the last highlighted option
+        if (this.asProps.visible) {
+          const index = this.prevHighlightedIndex;
+          this.handlers.highlightedIndex(index);
+        }
+      },
     };
   }
 
@@ -181,7 +198,8 @@ class RootSelect extends AbstractDropdown {
 
   getOptionProps(props, index) {
     const { value, highlightedIndex, focusSourceRef, size = 'm' } = this.asProps;
-    const highlighted = index === highlightedIndex && focusSourceRef.current === 'keyboard';
+    const highlighted =
+      index === highlightedIndex && focusSourceRef.current === 'keyboard' && !props.disabled;
     const selected = props.selected ?? isSelectedOption(value, props.value);
 
     return {