+++ stories/components/input-tags/advanced/examples/tags_with_addons.tsx
@@ -8,32 +8,32 @@ const Demo = () => {
   return (
     <Box>
       <InputTags size='m' state='normal'>
-        <InputTags.Tag editable={false} tabIndex={0}>
+        <InputTags.Tag editable={false}>
           <InputTags.Tag.Text>
             <InputTags.Tag.Addon>
               <Check />
             </InputTags.Tag.Addon>
-            Tag With text and addon
+            <InputTags.Tag.Text.Content>Tag With text and addon</InputTags.Tag.Text.Content>
           </InputTags.Tag.Text>
         </InputTags.Tag>
-        <InputTags.Tag editable={false} tabIndex={0}>
+        <InputTags.Tag editable={false}>
           <InputTags.Tag.Text>Tag with text and close</InputTags.Tag.Text>
           <InputTags.Tag.Close />
         </InputTags.Tag>
-        <InputTags.Tag editable={false} tabIndex={0}>
+        <InputTags.Tag editable={false}>
           <InputTags.Tag.Text>Tag with text</InputTags.Tag.Text>
         </InputTags.Tag>
-        <InputTags.Tag editable={true} tabIndex={0}>
+        <InputTags.Tag editable={true}>
           <InputTags.Tag.Text>
             <InputTags.Tag.Addon>
               <Edit />
             </InputTags.Tag.Addon>
-            Tag with addon text and close
+            <InputTags.Tag.Text.Content>Tag with addon text and close</InputTags.Tag.Text.Content>
           </InputTags.Tag.Text>
           <InputTags.Tag.Close />
         </InputTags.Tag>
 
-        <InputTags.Tag editable={true} tabIndex={0}>
+        <InputTags.Tag editable={true}>
           <InputTags.Tag.Text>
             <InputTags.Tag.Circle
               style={{
@@ -43,12 +43,12 @@ const Demo = () => {
             <InputTags.Tag.Addon>
               <Check />
             </InputTags.Tag.Addon>
-            Addon circle text and close
+            <InputTags.Tag.Text.Content>Addon circle text and close</InputTags.Tag.Text.Content>
           </InputTags.Tag.Text>
           <InputTags.Tag.Close />
         </InputTags.Tag>
 
-        <InputTags.Tag editable={false} tabIndex={0}>
+        <InputTags.Tag editable={false}>
           <InputTags.Tag.Text>
             <InputTags.Tag.Addon>
               <Edit />
@@ -56,7 +56,7 @@ const Demo = () => {
             <InputTags.Tag.Close />
           </InputTags.Tag.Text>
         </InputTags.Tag>
-        <InputTags.Tag editable={true} tabIndex={0}>
+        <InputTags.Tag editable={true}>
           <InputTags.Tag.Text>
             <InputTags.Tag.Addon>
               <Edit />
@@ -67,32 +67,32 @@ const Demo = () => {
       </InputTags>
 
       <InputTags size='l' state='normal'>
-        <InputTags.Tag editable={false} tabIndex={0}>
+        <InputTags.Tag editable={false}>
           <InputTags.Tag.Text>
             <InputTags.Tag.Addon>
               <Check />
             </InputTags.Tag.Addon>
-            Tag With text and addon
+            <InputTags.Tag.Text.Content>Tag With text and addon</InputTags.Tag.Text.Content>
           </InputTags.Tag.Text>
         </InputTags.Tag>
-        <InputTags.Tag editable={false} tabIndex={0}>
+        <InputTags.Tag editable={false}>
           <InputTags.Tag.Text>Tag with text and close</InputTags.Tag.Text>
           <InputTags.Tag.Close />
         </InputTags.Tag>
-        <InputTags.Tag editable={false} tabIndex={0}>
+        <InputTags.Tag editable={false}>
           <InputTags.Tag.Text>Tag with text</InputTags.Tag.Text>
         </InputTags.Tag>
-        <InputTags.Tag editable={true} tabIndex={0}>
+        <InputTags.Tag editable={true}>
           <InputTags.Tag.Text>
             <InputTags.Tag.Addon>
               <Edit />
             </InputTags.Tag.Addon>
-            Tag with addon text and close
+            <InputTags.Tag.Text.Content>Tag with addon text and close</InputTags.Tag.Text.Content>
           </InputTags.Tag.Text>
           <InputTags.Tag.Close />
         </InputTags.Tag>
 
-        <InputTags.Tag editable={true} tabIndex={0}>
+        <InputTags.Tag editable={true}>
           <InputTags.Tag.Text>
             <InputTags.Tag.Circle
               style={{
@@ -102,12 +102,12 @@ const Demo = () => {
             <InputTags.Tag.Addon>
               <Check />
             </InputTags.Tag.Addon>
-            Addon circle text and close
+            <InputTags.Tag.Text.Content>Addon circle text and close</InputTags.Tag.Text.Content>
           </InputTags.Tag.Text>
           <InputTags.Tag.Close />
         </InputTags.Tag>
 
-        <InputTags.Tag editable={false} tabIndex={0}>
+        <InputTags.Tag editable={false}>
           <InputTags.Tag.Text>
             <InputTags.Tag.Addon>
               <Edit />
@@ -115,7 +115,7 @@ const Demo = () => {
             <InputTags.Tag.Close />
           </InputTags.Tag.Text>
         </InputTags.Tag>
-        <InputTags.Tag editable={true} tabIndex={0}>
+        <InputTags.Tag editable={true}>
           <InputTags.Tag.Text>
             <InputTags.Tag.Addon>
               <Edit />
@@ -126,32 +126,32 @@ const Demo = () => {
       </InputTags>
 
       <InputTags size='m' state='invalid'>
-        <InputTags.Tag editable={false} tabIndex={0}>
+        <InputTags.Tag editable={false}>
           <InputTags.Tag.Text>
             <InputTags.Tag.Addon>
               <Check />
             </InputTags.Tag.Addon>
-            Tag With text and addon
+            <InputTags.Tag.Text.Content>Tag With text and addon</InputTags.Tag.Text.Content>
           </InputTags.Tag.Text>
         </InputTags.Tag>
-        <InputTags.Tag editable={false} tabIndex={0}>
+        <InputTags.Tag editable={false}>
           <InputTags.Tag.Text>Tag with text and close</InputTags.Tag.Text>
           <InputTags.Tag.Close />
         </InputTags.Tag>
-        <InputTags.Tag editable={false} tabIndex={0}>
+        <InputTags.Tag editable={false}>
           <InputTags.Tag.Text>Tag with text</InputTags.Tag.Text>
         </InputTags.Tag>
-        <InputTags.Tag editable={true} tabIndex={0}>
+        <InputTags.Tag editable={true}>
           <InputTags.Tag.Text>
             <InputTags.Tag.Addon>
               <Edit />
             </InputTags.Tag.Addon>
-            Tag with addon text and close
+            <InputTags.Tag.Text.Content>Tag with addon text and close</InputTags.Tag.Text.Content>
           </InputTags.Tag.Text>
           <InputTags.Tag.Close />
         </InputTags.Tag>
 
-        <InputTags.Tag editable={true} tabIndex={0}>
+        <InputTags.Tag editable={true}>
           <InputTags.Tag.Text>
             <InputTags.Tag.Circle
               style={{
@@ -161,12 +161,12 @@ const Demo = () => {
             <InputTags.Tag.Addon>
               <Check />
             </InputTags.Tag.Addon>
-            Addon circle text and close
+            <InputTags.Tag.Text.Content>Addon circle text and close</InputTags.Tag.Text.Content>
           </InputTags.Tag.Text>
           <InputTags.Tag.Close />
         </InputTags.Tag>
 
-        <InputTags.Tag editable={false} tabIndex={0}>
+        <InputTags.Tag editable={false}>
           <InputTags.Tag.Text>
             <InputTags.Tag.Addon>
               <Edit />
@@ -174,7 +174,7 @@ const Demo = () => {
             <InputTags.Tag.Close />
           </InputTags.Tag.Text>
         </InputTags.Tag>
-        <InputTags.Tag editable={true} tabIndex={0}>
+        <InputTags.Tag editable={true}>
           <InputTags.Tag.Text>
             <InputTags.Tag.Addon>
               <Edit />
@@ -185,32 +185,32 @@ const Demo = () => {
       </InputTags>
 
       <InputTags size='l' state='valid'>
-        <InputTags.Tag editable={false} tabIndex={0}>
+        <InputTags.Tag editable={false}>
           <InputTags.Tag.Text>
             <InputTags.Tag.Addon>
               <Check />
             </InputTags.Tag.Addon>
-            Tag With text and addon
+            <InputTags.Tag.Text.Content>Tag With text and addon</InputTags.Tag.Text.Content>
           </InputTags.Tag.Text>
         </InputTags.Tag>
-        <InputTags.Tag editable={false} tabIndex={0}>
+        <InputTags.Tag editable={false}>
           <InputTags.Tag.Text>Tag with text and close</InputTags.Tag.Text>
           <InputTags.Tag.Close />
         </InputTags.Tag>
-        <InputTags.Tag editable={false} tabIndex={0}>
+        <InputTags.Tag editable={false}>
           <InputTags.Tag.Text>Tag with text</InputTags.Tag.Text>
         </InputTags.Tag>
-        <InputTags.Tag editable={true} tabIndex={0}>
+        <InputTags.Tag editable={true}>
           <InputTags.Tag.Text>
             <InputTags.Tag.Addon>
               <Edit />
             </InputTags.Tag.Addon>
-            Tag with addon text and close
+            <InputTags.Tag.Text.Content>Tag with addon text and close</InputTags.Tag.Text.Content>
           </InputTags.Tag.Text>
           <InputTags.Tag.Close />
         </InputTags.Tag>
 
-        <InputTags.Tag editable={true} tabIndex={0}>
+        <InputTags.Tag editable={true}>
           <InputTags.Tag.Text>
             <InputTags.Tag.Circle
               style={{
@@ -220,12 +220,12 @@ const Demo = () => {
             <InputTags.Tag.Addon>
               <Check />
             </InputTags.Tag.Addon>
-            Addon circle text and close
+            <InputTags.Tag.Text.Content>Addon circle text and close</InputTags.Tag.Text.Content>
           </InputTags.Tag.Text>
           <InputTags.Tag.Close />
         </InputTags.Tag>
 
-        <InputTags.Tag editable={false} tabIndex={0}>
+        <InputTags.Tag editable={false}>
           <InputTags.Tag.Text>
             <InputTags.Tag.Addon>
               <Edit />
@@ -233,7 +233,7 @@ const Demo = () => {
             <InputTags.Tag.Close />
           </InputTags.Tag.Text>
         </InputTags.Tag>
-        <InputTags.Tag editable={true} tabIndex={0}>
+        <InputTags.Tag editable={true}>
           <InputTags.Tag.Text>
             <InputTags.Tag.Addon>
               <Edit />