+++ stories/patterns/filters/filter-cp-cd-cpc/docs/filter-cp-cd-cpc.stories.tsx
@@ -1,0 +1,13 @@
+import type { Meta, StoryObj } from '@storybook/react';
+
+import ClickPotentialCompetitiveDensityCPCExample from './examples/basic-example';
+
+const meta: Meta = {
+  title: 'Patterns/Filters/ClickPotentialCompetitiveDensityCPC',
+};
+
+export default meta;
+
+export const ClickPotentialCompetitiveDensityCPC: StoryObj = {
+  render: ClickPotentialCompetitiveDensityCPCExample,
+};